build(grpc): Use FetchContent to clone GRPC src
Instead of using the bash script to download GRPC,
we use CMake's FetchContent to download the GRPC
source-code instead.

This download is faster and shows progress.

Instead of deleting the folder, we now cache it,
to make building/testing muuch faster, especially
since the @nqminds office has awful internet connection.

However, the GRPC source and build folders are
now subfolders of the GRPC folder, so we
can clear it easily.

We still use the bash script to compile GRPC, however,
since @mereacre was having issues with cross-compilation.