Upgrade OpenSSL to v3.0.0
I was bored while waiting for the .debs to compile while debugging
https://github.com/nqminds/EDGESec/pull/40,
so I upgraded OpenSSL to v3.0.0 so we're no longer using a beta
version.

Main change is that it now installs to /lib64 by default now,
so we need to use --libdir=lib to go back to the same behaviour.