build: Check for glibc version in library data
On my computer, with gcc-11, there is no glibc filename
that contains the version number of glibc.

It's possible to easily find out the glibc by compiling and
running a program, however, this won't work without emulation
when cross-compiling.

Because of this, instead we use GNU strings to search for all
strings in the glibc library.

The library lists all the versions it supports in the format
GLIBC_x.y.z, so all we have to do is some simple string opts
to find the latest version, then to return that version.