build: Remove redundant glibc version check
The GLIBC version we're checking and throwing an error for is glibc
version 2.2.1, which came out in 2001.

Our CMake file now enables:
CMAKE_C_STANDARD 11 and CMAKE_C_STANDARD_REQUIRED on

This means glib 2.16 should be the minimum version of GLIBC that is
used, as that's the first version that came out with C11 support.