build(openwrt): use single toolchain location
Previously, the toolchain was being built in the CMAKE_BINARY_DIR.
However, libpcap uses the command, `check_include_file()`, which
creates a new CMake project in a temporary folder to test C/C++ features.

This causes the OpenWRT SDK to be constantly downloaded multiple times,
taking a lot of time.

Instead, I've changed the OpenWRT toolchain file to download and extract
the SDK next to the toolchain file. This should let the SDK be cached.