build: only build static or shared sqlite lib
Only build the static or the shared version of sqlite.

Building sqlite is one of the slowest parts of the build process.
We can cut down on the time the sqlite build takes by only
building the shared library version or the static library
versioning, depending on which version we need.

Please be aware, that due to multiprocessing, this changed
will only cut down speed on low processor count computers,
e.g. CI servers. Desktops/laptops probably won't be affected.