build: download uthash and utarray from upstream
Instead of maintaining our own fork of uthash and utarray,
we can download the upstream version.

This means we have less code to test, increasing code
coverage. It also makes it easier to update to a newer
version in the future.

The only changes we had made to our version was to use
src/util/allocs.h