ci: use buildDir from CMake presets
Currently, we're hardcoding the build directory `./build` when
building in CI, as the `cmake --install` command does not
support loading the build directory from a cmake-preset.

However, when running OpenWRT tests, the cmake-preset sets a
LD_LIBRARY_PATH that points to the build directory, so we
can no longer use `./build`.