build: use eloops from hostapd source code
The eloops.c/.h file from the 2.10 HostAPD source code contains
FreeBSD support. Instead of copying the file into our git repo,
we can instead download it, and manually apply patches to the file.

This should make it much easier to upgrade to v2.11 of HostAPD,
or to take in any other upstream improvements to eloop.