fix: fix if_nametoindex implicit declaration
The `#include <net/if.h>` where if_nametoindex was defined was not
included.

This doesn't cause any issues in Linux 32-bit, since there the implicit
`int` and actual pointer parameter are the same size.

However, on Linux 64-bit, this means that the compiler sees us passing
a 64-bit pointer to a 32-bit int, so who knows what will happen.