fix(dnsmasq): truncate interface names/prefix
When creating new interfaces from the wifi_interface + vlanid,
ensure that we always keep the full vlanid in the new interface name,
by truncating the wifi_interface/interface_prefix to the appropriate
number of bytes.

Since the max vlanid is 4094, and interface names are usually limited
to 16 bytes, we need to limit the interface_prefixes to 11 bytes.

Co-authored-by: Alexandru Mereacre <mereacre@gmail.com>