feat(os): add copy_argv() function
C programs such as run_command() accept argv as an `char *`, not
a `const char *`. This is because the `main()` function of programs
is explicitly allowed to modify the data in `argv()`.

This can cause issues, since we declare arrays using string literals,
(e.g. `char * argv[] = {"hello", "world", NULL};`), and modifying
string literals is undefined behaviour.

To get around this, we can just make a mofiiable copy of argv before
passing it to run_command().

There is a warning we can enable that will check for this
(`-Wwrite-strings`), but it creates 100s of warnings throughout our
code-base, so it's something we should slowly fix.