refactor(radius_client): fix ptr to enum cast
Casting a pointer to an int is not well defined.
However, casting a pointer to a uintptr_t is perfectly legal.
Then, casting a uintptr_t to a small int/enum is also perfectly legal
(although it may result in a loss of data).

The only issues is that uintptr_t might not exist on every platform,
but it probably exists on all the big platforms. I'm not sure if will
exist on Morello though, since Morello pointers are a bit odd.