test(sqlite_header): improve sqlite_header tests
The test_sqlite_header test is intermittently failing in
GitHub Actions.

I'm not 100% sure why this the test is failing.
One thing I did notice is that CMocka performs really poorly
in multi-threaded applications.

As recommended in the CMocka docs, I've added the CMOCKA_TEST_ABORT=1
environment variable so that CMocka automatically `aborts()` if a test
fails, see https://api.cmocka.org/#main-threads

I've also removed all of the `assert_*()` macros from the non-main
threads, since CMocka doesn't properly handle these. Instead, we
pass an error_message string to the main thread, and the main thread
then calls `asssert_string_equal()` on it.

This might fix https://github.com/nqminds/edgesec/issues/252,
but even if it doesn't, hopefully this will improve the logs and make
it easier to debug in the future.