fix(sockctl): log when deleting tmp folders
There's a very tiny chance that we'll misidentify a folder
as a mkdtemp folder created by `create_tmp_domain_socket_path()`.

For this case, I've added a log statement. It's only a log_debug(),
since we expect that if this ever does misidentify a folder, rmdir()
will throw an ENOTEMPTY (dir not empty) error, which will print
an error log message.