build: error CMake if C11 threads.h does not work
The C11 threads.h library is supported in almost every modern C11
compiler. However, since it's technically optional, we should
explicitly check if it's supported, so that we can print a nicer
error message.

Technically, the way you are supposed to do this is by checking
for __STDC_NO_THREADS__, but that's not easy to do in CMake.

See https://en.cppreference.com/w/c/thread