build(protobuf-c): pass host CMAKE vars to build
Pass the following variables from the host/root CMake project
to the protobuf-c CMake ExternalProject when configuring,
for improve cross-compiling:
  - CMAKE_LIBRARY_ARCHITECTURE
  - CMAKE_SYSTEM_NAME
  - CMAKE_SYSTEM_PROCESSOR
  - CMAKE_C_COMPILER

When using a CMAKE_TOOLCHAIN_FILE
(e.g. when cross-compiling using an OpenWRT SDK preset),
this does nothing. However, a toolchain file is not used when
cross-compiling for a .deb package when using pdebuild/debuild.