test: use AddCMockaTest to simplify cmake files
Use the AddCMockaTest.cmake script to simplify the test CMake
build scripts.

By default, all `LINK_LIBRARIES` are linked privately by default,
since the tests are executables and are not expected to be linked
to by other CMake targets.

There is a LINK_OPTIONS variable that the script uses, but confusingly,
it is not actually passed to CMake's `LINK_OPTIONS`.
It's instead passed to
[`LINK_FLAGS`][1]

[1]: https://cmake.org/cmake/help/latest/prop_tgt/LINK_FLAGS.html