build: support non-makefile CMake generators
Currently, all of our ExternalProjects using make are using '$(MAKE)'
so that their builds are run in the same make invocation as the
main repo.

(e.g. this means things like multiprocessing get shared with all
 ExternalProjects).

However, this does not work when the main project is using a
non-Makefile generator.