build: support using Ninja with CMake
Ninja 🥷 is much faster than GNU Make, and is now the default
CMake generator for OpenWRT.

It's simplicity is the reason why it's so fast, but it does have the
following issue:

> Underspecified build dependencies will result in incorrect builds.

To fix this, we need to manually declare the library files for
our ExternalProject libraries, using `BUILD_BYPRODUCTS`.

Technically, all of the library files are actually generated by
the ExternalProject INSTALL step, not the BUILD step, but we only
need to add these rules to make Ninja happy, so it's fine.

In the future, if CMake adds an `INSTALL_BYPRODUCTS` option,
we should use that instead.

See https://gitlab.kitware.com/cmake/cmake/-/issues/23056