ci(build): use ubuntu-22.04 to run CI
ubuntu-latest is in the process of migrating to using ubuntu-22.04.
This means that it is effectively random whether our CI runner
uses ubuntu-20.04 or ubuntu-22.04, which is a problem, as due to a
version mistmatch in ubuntu-20.04's default clang and default LLVM
version, we currently manually specify clang v10.

ubuntu-22.04 uses the same version for both clang and LLVM by
default, so using that runner means we can remove our workaround.
We should also get better error messages and more optimized code.