build: load `--host` from CMAKE_C_COMPILER_TARGET
Load the `--host` parameter for autoconf from the
CMAKE_C_COMPILER_TARGET value.

This is the variable that is recommended to be used, see
https://cmake.org/cmake/help/latest/variable/CMAKE_LANG_COMPILER_TARGET.html

As a backup, we still load from CMAKE_LIBRARY_ARCHITECTURE
if CMAKE_C_COMPILER_TARGET is empty, as when cross-compiling
a debian using debuild, CMAKE_C_COMPILER_TARGET is not set.