refactor(ap_service): fix run_ap callback type
In ISO C, casting from a function pointer to a void pointer is
undefined. Some platforms (e.g. z/OS or MS-DOS) have different sized
pointers for functions and data.

To get around this, we can instead pass run_ap() a pointer to a
struct that has a pointer to a function.