build(src): compile with -Wpedantic
Compile with the `-Wpedandtic` flag in the edgesec `src/` directory.
`-Wpedandtic` checks for a lot more C errors that are often allowed
by the GCC/Clang compilers, but are technically incorrect in ISO C.

However, it's worth enabling, as since we are compiling for CheriBSD,
we want to follow standard C as much as possible.

We need to disable the `-Wpedandtic` flag when compiling `nl`, however.
This is because nl includes linux/netlink, which includes some 0-length
arrays, which are invalid in ISO C.