refactor(eloop): add `edge_` prefix to funcs
Add the `edge_` namespace prefix to all eloop functions.

We're having difficulties with using the libeap library, because it
also uses eloop, but a different version of eloop.
This is causing linking errors.

Unfortunately, we're not using C++, so we can't just wrap everything
in a namespace, so we have to manually namespace the functions by
adding our own custom prefix.

Style changes will be in a seperate commit to make viewing the
`git diff` a bit easier.