build(cmocka): make upstream cmocka mandatory
Currently, using cmocka v1.1.5 worked with our current code-base,
as long as we weren't running tests in Morello or CheriBSD, as
we were using syntax compatbile with both CMocka v1.1.5 and the current
upstream.

However, due to breaking changes on the `main` branch, we now must
use the upstream CMocka syntax for Morello and CheriBSD compatibility.