refactor: inline os_zalloc for internal linkage
Change os_zalloc to use `static inline` so that it used internal
linkage.

The libeap library also has it's own os_zalloc, which is causing
linking errors. We can easily avoid this by inlining the edgesec
version of `os_zalloc`. The function is literally just one-line of
code, so inlining it should basically change nothing.

Co-authored-by: Alexandru Mereacre <mereacre@gmail.com>