build(uci): download uci archive, not `git`
Download uci (and libubox) as a tarball, instead of using a
`git clone`.

Both `uci` and `libubox` are hosted on the [OpenWRT git server][1],
which isn't very fast.

Downloading a tarball is quite a bit faster. Additionally, it means
that we can easily store it in a custom `DOWNLOAD_DIR` and easily
cache it, to prevent mutliple re-downloads.

Finally, doing `rm -r` on a git repo fails
(you need `-f` to force delete it). Deleting an uncompressed tarball
is much easier.

A few issues:
  - In CMake v3.22.1, `;` is a list seperator character.
    Because GitWeb uses `;` in the URLs, we need to escape them in
    CMake using `\\\;` (double-escaped because CMake has some
    [interesting list unescaping issues][2]).

    We also can't use [`$<SEMICOLON>`][3], because the URL parameter
    of `ExternalProject_Add()` doesn't yet support generator
    expressions.
  - `git.openwrt.org` is currently running git v2.30.2.
    When it updates to git v2.38.0+, the default gzip compression
    algorithm will change, which may cause the hash to change,
    see [git/git@4f4be00d302bc52d0d9d5a3d4738bb525066c710][4]

[1]: https://git.openwrt.org/
[2]: https://gitlab.kitware.com/cmake/cmake/-/issues/20317
[3]: https://cmake.org/cmake/help/v3.22/manual/cmake-generator-expressions.7.html#escaped-characters
[4]: https://github.com/git/git/commit/4f4be00d302bc52d0d9d5a3d4738bb525066c710