feat(common): add `bin_clear_free` function
Add `bin_clear_free()`, which is used to set memory to `'\0'` before
free()-ing it, to avoid leaking sensitive information.

Required by code taken from `hostap`.

Unfortunately, the memset() might still be optimized out by a smart
compiler. C11 has `memset_s`, but it's part of Annex K, which means
basically nobody except Microsoft supports it, so there's not much
point if us trying to use it.

C23 has `memset_explicit`, but C23 is not out yet, so we can't use it.

Alternatively, we could use the non-standard [`explicit_bzero`][1]
function. It's available in glibc, and most BSDs. But, we'd probably
then need to:
  - a) add CMake code to call [`check_symbol_exists`][2]
  - b) add it to a `.c` file, not a `.h` file, to set the appropriate
       glibc flags.

Co-authored-by: Alexandru Mereacre <mereacre@gmail.com>
[1]: https://manpages.ubuntu.com/manpages/jammy/man3/bzero.3.html
[2]: https://cmake.org/cmake/help/latest/module/CheckSymbolExists.html