refactor(allocs): add `__must_check` attribute
Add a `__must_check` attribute to that tells Clang/GCC that
the return code for a function must be used.

This attribute may be used by code taken from the `hostap` project.

The original implementation from hostap was GCC specific, so I modified
it to work with Clang as well.

Co-authored-by: Alexandru Mereacre <mereacre@gmail.com>