style(common): add `__maybe_unused` attribute
Add the `__maybe_unused` attribute, which prevents warnings when a
variable is unused
(e.g. if we use the variable only if a given `#ifdef X` is set).

In the edgesec project, we normally use a cast to `(void)` to do the
same thing, e.g. `(void)my_var;`, but code taken from hostap
uses `__maybe_unused`.

This is equivalent to the C23 attribute [`[[maybe_unused]]`][1].

[1]: https://en.cppreference.com/w/c/language/attributes/maybe_unused