refactor: add header file for compiler attributes
Add `./src/utils/attributes.h` header file for compiler attributes.

For example, `__attribute__((unused))` is supported on GCC and Clang,
but on other compilers, this is not supported.

Most of these are already in `src/radius/common.h`, so we just need to
move them into a new file to use them in the rest of the edgesec code.