test(edgesec): stop eloop threads gracefully
Currently, the eloop threads are never actually stopped.
We call `edge_eloop_terminate()`, but this is not threadsafe.

To fix this, we can use a threadsafe atomic from `<stdatomic.h>`,
which is constantly checked by eloop, which then automatically closes
the eloop when the atomic is set.

This also allows us to call `thrd_join()` on the threads and check
their return code.