test(edgesec): store error msg from ap thread
Currently, in `test_edgesec`, if the AP thread has an error,
`fail_msg()` is called, which normally fails the CMocka test with the
given error message. However, [CMocka **is not thread safe**][1].

Instead, we can make the AP Thread store the error message in a
buffer, and call `thrd_exit(EXIT_FAILURE)` to exit the thread with
an error exit code.

We can then check for this thread exitcode later on, when we call
`thrd_join()`, e.g. in `teardown_edgesec_test()`.

[1]: https://api.cmocka.org/