perf(os): return early from `is_proc_app()`
Return early from `is_proc_app()` without checking for
`is_string_in_cmdline_file()` if anything `proc_name` is in the
basename of the realpath of the process.

I've also scoped a bunch of the code, in order to reduce the amount
of buffers stored in the stack at once.