test(edgesec): speed up `test_edgesec`
Instead of using `writeread_domain_data_str`, and waiting for it
to timeout due to a lack of a response, we can just manually create
a socket and use `write_domain_data_s()`, to avoid even trying
to read a response.

This causes our test to be much faster (5x faster).