test(edgesec): test "ATTACH" to ap server
Currently, only the OpenWRT test presets of OpenWRT sends an "ATTACH" command to the AP server, which is compiled with GCC.

This causes an issue with code-coverage, because this means that this branch is only covered by code compiled/tested with
GCC, and not via LLVM/Clang. Testing this branch in Clang
should fix this code-coverage issue.