build: remove dependency on C++
EDGESec only uses C code, so we don't need to get CMake to search for
a C++ compiler (this occasionally causes issues with Clang).

OpenSSL still expects a CXX value, but it's not used for the OpenSSL
build we compile, so we can just set it to an executable that does
not exists.