test(eloop): test eloop_destroy()
The code-coverage on eloop_destroy() is currently inconsistent.

This is because there is a branch that only runs
`if (timeout->time.usec < now.usec) {`, where `now` is created from
`os_get_reltime()`.

I've made a test that checks for `eloop_destroy()` using a timeout
that has a `usec` of `0`.
Because there's a 1-in-a-million chance that `now.usec` is 0,
I'm running this test twice with a 1 microsecond delay, so that this
branch should always be hit.