fix(os): fix `edge_os_get_reltime()`
Fix `edge_os_get_reltime()` so that it correctly gets the relative
time.

It does this by using the [POSIX `CLOCK_MONOTONIC`][1], which defines
a monotonic clock that never goes backwards.

Because `CLOCK_MONOTONIC` is not supported on all POSIX platforms,
as backup we can use `CLOCK_REALTIME`, which is the same behavior as
edge_os_get_time() (aka the old behavior).

[1]: https://pubs.opengroup.org/onlinepubs/000095399/functions/clock_getres.html
Fixes: https://github.com/nqminds/edgesec/issues/408