test(ap_service): fix writeread_domain_data_str mock
Fix a memory leak in the `__wrap_writeread_domain_data_str()`
function, that mocks `writeread_domain_data_str()`.

If the return code of `writeread_domain_data_str()` is non-0,
the caller will **NOT** deallocate the reply pointer.

Because of that, we must only allocate the reply string if the
return code is 0.