refactor(dnsmasq): make signal_dhcp_process const
Make the `dhcp_bin_path` parameter to signal_dhcp_process() `const`.

Because we call `basename()` on this value, which may change the input
string, we need to first make a temporary copy of the string into a
buffer.

This fixes the segmentation fault in `test_signal_dhcp_process()`
test on FreeBSD.