test(mdns_service): fix uninitialized data in test
The `struct mdns_context context` data was not initialized when it was
used in `test_run_mdns` and `test_close_mdns`.

On Linux, this wasn't a problem, since the memory just happened to be
initialized to NULL, but on FreeBSD, the memory was filled with
garbage, which causes a Segmentation fault when trying to run
`init_reflections()` on a garbage pointer.

Fixes the `test_mdns_service` test on FreeBSD.